app/api/pyproject.toml
[tool.poetry]
name = "api"
version = "2.2.5"
description = "Web app for Carrot-Mapper"
authors = ["Sam Cox <sam.cox@nottingham.ac.uk>", "Philip Quinlan <philip.quinlan@nottingham.ac.uk>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.11"
django = "4.2.13"
openpyxl = "3.1.3"
django-select2 = "^8.1.2"
django-extra-views = "^0.14.0"
psycopg2 = "^2.9.9"
djangorestframework = "^3.15.2"
django-filter = "^24.2"
requests = "^2.31.0"
azure-common = "^1.1.28"
azure-storage-queue = "^12.10.0"
azure-storage-blob = "^12.20.0"
whitenoise = "^6.7.0"
gunicorn = "22.0.0"
graphviz = "^0.20.3"
drf-dynamic-fields = "^0.4.0"
django-cors-headers = "^4.4.0"
python-dotenv = "^1.0.1"
django-revproxy = "^0.12.0"
shared = {path = "../shared", develop = true}

[tool.poetry.group.test.dependencies]
pytest-django = "^4.8.0"


[tool.poetry.group.dev.dependencies]
django-stubs = "^5.0.2"
djangorestframework-stubs = "^3.15.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"
DJANGO_SETTINGS_MODULE = "test.settings"
